+++ domain/src/main/java/br/com/itau/desafioseguros/domain/entities/InsuranceProduct.java
@@ -4,6 +4,8 @@ import br.com.itau.desafioseguros.domain.exceptions.*;
 import br.com.itau.desafioseguros.domain.valueobjects.InsuranceProductCategory;
 import br.com.itau.desafioseguros.domain.valueobjects.InsuranceProductId;
 
+import java.math.BigDecimal;
+
 public class InsuranceProduct {
 
     private final InsuranceProductId insuranceProductId;
@@ -12,15 +14,15 @@ public class InsuranceProduct {
 
     private final InsuranceProductCategory category;
 
-    private final Float basePrice;
+    private final BigDecimal basePrice;
 
-    private final Float tariffedPrice;
+    private final BigDecimal tariffedPrice;
 
     private InsuranceProduct(InsuranceProductId insuranceProductId,
                              String name,
                              InsuranceProductCategory category,
-                             Float basePrice,
-                             Float tariffedPrice) {
+                             BigDecimal basePrice,
+                             BigDecimal tariffedPrice) {
         this.insuranceProductId = insuranceProductId;
         this.name = name;
         this.category = category;
@@ -31,8 +33,8 @@ public class InsuranceProduct {
     public static InsuranceProduct create(InsuranceProductId insuranceProductId,
                                           String name,
                                           InsuranceProductCategory category,
-                                          Float basePrice,
-                                          Float tariffedPrice) {
+                                          BigDecimal basePrice,
+                                          BigDecimal tariffedPrice) {
         if (insuranceProductId == null)
             throw new InsuranceProductIdNullException();
 
@@ -67,11 +69,11 @@ public class InsuranceProduct {
         return category;
     }
 
-    public Float getBasePrice() {
+    public BigDecimal getBasePrice() {
         return basePrice;
     }
 
-    public Float getTariffedPrice() {
+    public BigDecimal getTariffedPrice() {
         return tariffedPrice;
     }
 }
